+++ tracks.h
@@ -8,28 +8,26 @@
 class Tracks
 {
   public:
-
-  bool begin( int, int, int );
-  void setThrottle( int );
-  void setThrottleScale( int, int );
-  void setSteering( int );
-  void setSteeringScale( int, int );
-  void setGear( int );
-  void setGearScale( int, int );
-  void showDebug();
+    bool begin( uint8_t, uint8_t, uint8_t );
+    void setThrottle( uint8_t );
+    void setThrottleScale( uint8_t, uint8_t );
+    void setSteering( uint8_t );
+    void setSteeringScale( uint8_t, uint8_t );
+    void setGear( uint8_t );
+    void setGearScale( uint8_t, uint8_t );
+    void showDebug();
 
   private:
-
-  int m_throttleScaleMin = 0;
-  int m_throttleScaleMax = 180;
-  int m_steeringScaleMin = 0;
-  int m_steeringScaleMax = 180;
-  int m_gearScaleMin = 0;
-  int m_gearScaleMax = 180;
+    uint8_t m_throttleScaleMin = 0;
+    uint8_t m_throttleScaleMax = 180;
+    uint8_t m_steeringScaleMin = 0;
+    uint8_t m_steeringScaleMax = 180;
+    uint8_t m_gearScaleMin = 0;
+    uint8_t m_gearScaleMax = 180;
   
-  myPWMServo * throttle;
-  myPWMServo * steering;
-  myPWMServo * gear;
+    myPWMServo * throttle;
+    myPWMServo * steering;
+    myPWMServo * gear;
 };
 
 extern Tracks tracks;
